Move vgem workaround to its own .so and use on all platforms
Fixes #1450
