+++ targets/xiwi
@@ -17,6 +17,9 @@ if [ ! -f "/sys/class/tty/tty0/active" ]; then
     compile freon '-ldl' so
 fi
 
+# We need xorg workarounds on all platforms
+compile xorg '-ldl' so
+
 ltspackages=''
 # On non-ARM precise, install lts-trusty xorg server for compatibility with xorg
 # if kernel version != 3.4 (lts-trusty mesa requires version >=3.6)
